requirements.txt
absl-py==2.1.0
accelerate==1.6.0
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.0.1
braindecode==0.8.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
decorator==5.1.1
docstring-inheritance==2.2.1
docutils==0.21.2
einops==0.8.0
EMD-signal==1.6.4
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
frozendict==2.4.6
fsspec==2025.3.2
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.69.2
grpcio==1.68.0
h5py==3.12.1
hmmlearn==0.3.3
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.30.1
idna==3.10
imagesize==1.4.1
indic-nlp-library==0.92
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mne==1.8.0
Morfessor==2.0.6
mpmath==1.3.0
msgpack==1.1.0
multitasking==0.0.11
murmurhash==1.0.11
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.14.2
nbformat==5.10.4
nbmerge==0.0.4
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pathos==0.3.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.28.3
psutil==7.0.0
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
skorch==1.0.0
slack_sdk==3.35.0
smart-open==7.0.5
smmap==5.0.1
snowballstemmer==2.2.0
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.4.8
streamlit==1.40.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.18.0.post0
tf_keras==2.18.0
thinc==9.1.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
vaderSentiment==3.3.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.0
yfinance==0.2.52
